# [Performance] Make improvements to key validation