… tests/Integration/SettingTest.php