Request: `{{url}}/channels?include=meeting-schedule,soapbox,subscription,channel-settings,user-channel-settings,meeting-rating-question`
Number of channels loaded: 12

For these tests I added timers to the `KeyValidator` and counted the number of times it was called as well as the runtime.

##### Test 1: Ran on master
```
Called: 240 times
Total time: 179.96655161 ms
Average time: 0.74986063170833 ms
```

##### Test 2: Ran after commit 1 - Move validation to store method
```
Called: 76 times
Total time: 52.97167935 ms
Average time: 0.69699578092105 ms
```

##### Test 3: Ran after commit 2 - Remove redudant validation when loading multiple keys
```
Called: 74 times
Total time: 44.04785123 ms
Average time: 0.59524123283784 ms
```

##### Test 4: Ran after commit 3 - Stop using Laravel's validator
```
Called: 74 times
Total time: 0.7055662 ms
Average time: 0.0095346783783784 ms
```

So, this PR shaved about 180 ms off the response time with my local dataset.